+++ wgpu-core/src/command/transfer.rs
@@ -21,7 +21,7 @@ use crate::{
         TextureInitTrackerAction,
     },
     resource::{
-        MissingBufferUsageError, MissingTextureUsageError, ParentDevice, RawResourceAccess,
+        Buffer, MissingBufferUsageError, MissingTextureUsageError, ParentDevice, RawResourceAccess,
         Texture, TextureErrorDimension,
     },
     snatch::SnatchGuard,
@@ -33,7 +33,7 @@ pub type TexelCopyBufferInfo = wgt::TexelCopyBufferInfo<BufferId>;
 pub type TexelCopyTextureInfo = wgt::TexelCopyTextureInfo<TextureId>;
 pub type CopyExternalImageDestInfo = wgt::CopyExternalImageDestInfo<TextureId>;
 
-#[derive(Clone, Copy, Debug)]
+#[derive(Clone, Copy, Debug, Eq, PartialEq)]
 pub enum CopySide {
     Source,
     Destination,
@@ -276,9 +276,11 @@ pub(crate) fn extract_texture_selector<T>(
 ///
 /// Copied with some modifications from WebGPU standard.
 ///
-/// If successful, returns a pair `(bytes, stride)`, where:
+/// If successful, returns a tuple `(bytes, stride, is_contiguous)`, where:
 /// - `bytes` is the number of buffer bytes required for this copy, and
 /// - `stride` number of bytes between array layers.
+/// - `is_contiguous` is true if the linear texture data does not have padding
+///   between rows or between images.
 ///
 /// [vltd]: https://gpuweb.github.io/gpuweb/#abstract-opdef-validating-linear-texture-data
 pub(crate) fn validate_linear_texture_data(
@@ -288,7 +290,7 @@ pub(crate) fn validate_linear_texture_data(
     buffer_size: BufferAddress,
     buffer_side: CopySide,
     copy_size: &Extent3d,
-) -> Result<(BufferAddress, BufferAddress), TransferError> {
+) -> Result<(BufferAddress, BufferAddress, bool), TransferError> {
     let wgt::BufferTextureCopyInfo {
         copy_width,
         copy_height,
@@ -303,14 +305,14 @@ pub(crate) fn validate_linear_texture_data(
         width_blocks: _,
         height_blocks,
 
-        row_bytes_dense: _,
+        row_bytes_dense,
         row_stride_bytes,
 
         image_stride_rows: _,
         image_stride_bytes,
 
         image_rows_dense: _,
-        image_bytes_dense: _,
+        image_bytes_dense,
 
         bytes_in_copy,
     } = layout.get_buffer_texture_copy_info(format, aspect, copy_size)?;
@@ -347,7 +349,10 @@ pub(crate) fn validate_linear_texture_data(
         });
     }
 
-    Ok((bytes_in_copy, image_stride_bytes))
+    let is_contiguous = (row_stride_bytes == row_bytes_dense || !requires_multiple_rows)
+        && (image_stride_bytes == image_bytes_dense || !requires_multiple_images);
+
+    Ok((bytes_in_copy, image_stride_bytes, is_contiguous))
 }
 
 /// Validate the source format of a texture copy.
@@ -733,6 +738,90 @@ fn handle_dst_texture_init(
     Ok(())
 }
 
+/// Handle initialization tracking for a transfer's source or destination buffer.
+///
+/// Ensures that the transfer will not read from uninitialized memory, and updates
+/// the initialization state information to reflect the transfer.
+fn handle_buffer_init(
+    cmd_buf_data: &mut CommandBufferMutable,
+    info: &TexelCopyBufferInfo,
+    buffer: &Arc<Buffer>,
+    direction: CopySide,
+    required_buffer_bytes_in_copy: BufferAddress,
+    is_contiguous: bool,
+) {
+    const ALIGN_SIZE: BufferAddress = wgt::COPY_BUFFER_ALIGNMENT;
+    const ALIGN_MASK: BufferAddress = wgt::COPY_BUFFER_ALIGNMENT - 1;
+
+    let start = info.layout.offset;
+    let end = info.layout.offset + required_buffer_bytes_in_copy;
+    if !is_contiguous || direction == CopySide::Source {
+        // If the transfer will read the buffer, then the whole region needs to
+        // be initialized.
+        //
+        // If the transfer will not write a contiguous region of the buffer,
+        // then we need to make sure the padding areas are initialized. For now,
+        // initialize the whole region, although this could be improved to
+        // initialize only the necessary parts if doing so is likely to be
+        // faster than initializing the whole thing.
+        //
+        // Adjust the start/end outwards to 4B alignment.
+        let aligned_start = start & !ALIGN_MASK;
+        let aligned_end = (end + ALIGN_MASK) & !ALIGN_MASK;
+        cmd_buf_data.buffer_memory_init_actions.extend(
+            buffer.initialization_status.read().create_action(
+                buffer,
+                aligned_start..aligned_end,
+                MemoryInitKind::NeedsInitializedMemory,
+            ),
+        );
+    } else {
+        // If the transfer will write a contiguous region of the buffer, then we
+        // don't need to initialize that region.
+        //
+        // However, if the start and end are not 4B aligned, we need to make
+        // sure that we don't end up trying to initialize non-4B-aligned regions
+        // later.
+        //
+        // Adjust the start/end inwards to 4B alignment, we will handle the
+        // first/last pieces differently.
+        let aligned_start = (start + ALIGN_MASK) & !ALIGN_MASK;
+        let aligned_end = end & !ALIGN_MASK;
+        if aligned_start != start {
+            cmd_buf_data.buffer_memory_init_actions.extend(
+                buffer.initialization_status.read().create_action(
+                    buffer,
+                    aligned_start - ALIGN_SIZE..aligned_start,
+                    MemoryInitKind::NeedsInitializedMemory,
+                ),
+            );
+        }
+        if aligned_start != aligned_end {
+            cmd_buf_data.buffer_memory_init_actions.extend(
+                buffer.initialization_status.read().create_action(
+                    buffer,
+                    aligned_start..aligned_end,
+                    MemoryInitKind::ImplicitlyInitialized,
+                ),
+            );
+        }
+        if aligned_end != end {
+            // It is possible that `aligned_end + ALIGN_SIZE > dst_buffer.size`,
+            // because `dst_buffer.size` is the user-requested size, not the
+            // final size of the buffer. The final size of the buffer is not
+            // readily available, but was rounded up to COPY_BUFFER_ALIGNMENT,
+            // so no overrun is possible.
+            cmd_buf_data.buffer_memory_init_actions.extend(
+                buffer.initialization_status.read().create_action(
+                    buffer,
+                    aligned_end..aligned_end + ALIGN_SIZE,
+                    MemoryInitKind::NeedsInitializedMemory,
+                ),
+            );
+        }
+    }
+}
+
 impl Global {
     pub fn command_encoder_copy_buffer_to_buffer(
         &self,
@@ -1004,7 +1093,7 @@ impl Global {
                 true, // alignment required for buffer offset
             )?;
 
-            let (required_buffer_bytes_in_copy, bytes_per_array_layer) =
+            let (required_buffer_bytes_in_copy, bytes_per_array_layer, is_contiguous) =
                 validate_linear_texture_data(
                     &source.layout,
                     dst_texture.desc.format,
@@ -1020,12 +1109,13 @@ impl Global {
                     .map_err(TransferError::from)?;
             }
 
-            cmd_buf_data.buffer_memory_init_actions.extend(
-                src_buffer.initialization_status.read().create_action(
-                    &src_buffer,
-                    source.layout.offset..(source.layout.offset + required_buffer_bytes_in_copy),
-                    MemoryInitKind::NeedsInitializedMemory,
-                ),
+            handle_buffer_init(
+                cmd_buf_data,
+                source,
+                &src_buffer,
+                CopySide::Source,
+                required_buffer_bytes_in_copy,
+                is_contiguous,
             );
 
             let regions = (0..array_layer_count)
@@ -1124,7 +1214,7 @@ impl Global {
                 true, // alignment required for buffer offset
             )?;
 
-            let (required_buffer_bytes_in_copy, bytes_per_array_layer) =
+            let (required_buffer_bytes_in_copy, bytes_per_array_layer, is_contiguous) =
                 validate_linear_texture_data(
                     &destination.layout,
                     src_texture.desc.format,
@@ -1180,13 +1270,13 @@ impl Global {
             let dst_barrier =
                 dst_pending.map(|pending| pending.into_hal(&dst_buffer, &snatch_guard));
 
-            cmd_buf_data.buffer_memory_init_actions.extend(
-                dst_buffer.initialization_status.read().create_action(
-                    &dst_buffer,
-                    destination.layout.offset
-                        ..(destination.layout.offset + required_buffer_bytes_in_copy),
-                    MemoryInitKind::ImplicitlyInitialized,
-                ),
+            handle_buffer_init(
+                cmd_buf_data,
+                destination,
+                &dst_buffer,
+                CopySide::Destination,
+                required_buffer_bytes_in_copy,
+                is_contiguous,
             );
 
             let regions = (0..array_layer_count)